backend/requirements.txt
# --- API & Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.5
python-multipart==0.0.9
pydantic==2.9.2

# --- DB (Postgres/Supabase) ---
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9

# --- Auth ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0

# --- AWS Bedrock / RAG ---
boto3==1.34.162
faiss-cpu==1.8.0
beautifulsoup4
lxml
pdfminer.six
tqdm
numpy==1.26.4
pypdf==5.0.1
requests==2.32.3
packaging>=23.2

# --- Offline Simulation (TF-IDF) ---
scikit-learn==1.5.1

# --- Convenience ---
python-dotenv==1.0.1
